requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.8.0
Authlib==1.4.1
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
databases==0.9.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.8
fonttools==4.55.8
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uv==0.5.26
uvicorn==0.34.0
zipp==3.21.0
